# api/swagger.yaml
openapi: '3.0.2'
info:
  title: FoREST
  version: '1.0'
servers:
  - url: https://backend.mdma.haveachin.de/
  - url: http://localhost:8080

# ENDPOINTS ########################################################################
paths:
  /mesh-nodes/{uuid}:
    parameters:
      - $ref: "#/components/parameters/UUID"
    get:
      tags:
        - Mesh-Nodes
      responses:
        200:
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMeshNode"
        400:
          description: Bad Request.
        404:
          description: Not Found.
        500:
          description: Internal Server Error.

    put:
      tags:
        - Mesh-Nodes
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutMeshNode"
      responses:
        200:
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMeshNode"
        400:
          description: Bad Request.
        404:
          description: Not Found.
        500:
          description: Internal Server Error.

    delete:
      tags:
        - Mesh-Nodes
      responses:
        204:
          description: OK.
        400:
          description: Bad Request.
        404:
          description: Not Found.
        500:
          description: Internal Server Error.
  /mesh-nodes:
    get:
      tags:
        - Mesh-Nodes
      responses:
        200:
          description: OK.
          content:
            application/json:
              schema:
                type: array
                minItems: 1
                items:
                  $ref: "#/components/schemas/GetMeshNode"
        500:
          description: Internal Server Error.

    post:
      tags:
        - Mesh-Nodes
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostMeshNode"
      responses:
        201:
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMeshNode"
        400:
          description: Bad Request.
        500:
          description: Internal Server Error.
  /mesh-nodes/{uuid}/data:
    parameters:
      - $ref: "#/components/parameters/UUID"
    post:
      tags:
        - Mesh-Nodes
        - Data
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostData"
      responses:
        201:
          description: Created.
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/GetSingleData"
        400:
          description: Bad Request.
        500:
          description: Internal Server Error.
  /mesh-nodes/{uuid}/data-list:
    parameters:
      - $ref: "#/components/parameters/UUID"
    post:
      tags:
        - Mesh-Nodes
        - Data
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/PostData"
      responses:
        201:
          description: Created.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetSingleData"
        400:
          description: Bad Request.
        500:
          description: Internal Server Error.
...
  /roles:
    get:
      tags:
        - Roles
      responses:
        200:
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetRole"
        500:
          description: Internal Server Error.
    post:
      tags:
        - Roles
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Role"
      responses:
        201:
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetRole"
        400:
          description: Bad Request.
        500:
          description: Internal Server Error.
  /roles/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags:
        - Roles
      responses:
        200:
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetRole"
        400:
          description: Bad Request.
        404:
          description: Not Found.
        500:
          description: Internal Server Error.
    put:
      tags:
        - Roles
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Role"
      responses:
        200:
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetRole"
        400:
          description: Bad Request.
        404:
          description: Not Found.
        500:
          description: Internal Server Error.

    delete:
      tags:
        - Roles
      responses:
        204:
          description: No Content.
        400:
          description: Bad Request.
        404:
          description: Not Found.
        500:
          description: Internal Server Error.
  /roles/permissions:
    get:
      tags:
        - Roles
      responses:
        200:
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Permission"
        500:
          description: Internal Server Error.
...
  /mesh-node-updates:
    get:
      tags:
        - Updates
      responses:
        200:
          description: OK.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetUpdate"
        500:
          description: Internal Server Error.

    post:
      tags:
        - Updates
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Update"
      responses:
        201:
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUpdate"
        400:
          description: Bad Request.
        500:
          description: Internal Server Error.
  /mesh-node-updates/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags:
        - Updates
      responses:
        200:
          description: OK.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUpdate"
        400:
          description: Bad Request.
        404:
          description: Not Found.
        500:
          description: Internal Server Error.
...
components:
  schemas:
    UUID:
      type: string
      example: "0cc56633-05ae-4cc3-8f71-801f429caeca"
    ID:
      type: integer
      example: 1
    ResourceWithUUID:
      type: object
      properties:
        uuid:
          $ref: "#/components/schemas/UUID"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ResourceWithID:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/ID"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Update:
      type: object
      properties:
        data:
          type: string
          example: UG9seWZvbiB6d2l0c2NoZXJuZCBhw59lbiBNw6R4Y2hlbnMgVsO2Z2VsIFLDvGJlbiwgSm9naHVydCB1bmQgUXVhcms=
        version:
          type: string
          example: 1.0.0
    GetUpdate:
      allOf:
        - $ref: "#/components/schemas/Update"
      properties:
        id:
          $ref: "#/components/schemas/ID"
        createdAt:
          type: string
          format: date-time
...
    Role:
      type: object
      properties:
        name:
          type: string
          example: "admin"
        permissions:
          type: array
          items:
            type: string
            example: mesh_node_create
    GetRole:
      allOf:
        - $ref: "#/components/schemas/ResourceWithID"
        - $ref: "#/components/schemas/Role"
    Permission:
      type: string
      example: sensor_data_read
...
    MeshNode:
      type: object
      properties:
        updateId:
          $ref: "#/components/schemas/ID"
        latitude:
          type: number
          example: 49.127327
        longitude:
          type: number
          example: 9.264715
    GetMeshNode:
      allOf:
        - $ref: "#/components/schemas/ResourceWithUUID"
        - $ref: "#/components/schemas/MeshNode"
    PostMeshNode:
      allOf:
        - type: object
          properties:
            uuid:
              $ref: "#/components/schemas/UUID"
        - $ref: "#/components/schemas/MeshNode"
    PutMeshNode:
      $ref: "#/components/schemas/MeshNode"
...
    SingleData:
      type: object
      properties:
        measuredAt:
          type: string
          format: date-time
        type:
          type: string
          example: "temperature"
        value:
          type: string
          example: "23.23423"
    PostData:
      $ref: "#/components/schemas/SingleData"
    GetSingleData:
      allOf:
        - $ref: "#/components/schemas/SingleData"
      properties:
        id:
          $ref: "#/components/schemas/ID"
        createdAt:
          type: string
          format: date-time
  parameters:
    UUID:
      name: uuid
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
    ID:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ID"